refactor(imap): extract imap_probe to dialog_helpers (cycle 53)
