+++ src/external_imap/dialog.rs
@@ -1,51 +1,14 @@
 //! Dialogues IMAP hand-rollés (extrait de mod.rs).
 
 use chrono::Utc;
-use openssl::ssl::{SslConnector, SslMethod, SslStream};
+use openssl::ssl::SslStream;
 use std::io::Write;
-use std::net::{TcpStream, ToSocketAddrs};
-use std::time::Duration;
+use std::net::TcpStream;
 
 use super::parser::{
     escape_imap, parse_capabilities, parse_fetch_headers, parse_list_folders, parse_uid_search,
     read_line_from_stream, read_until_tag_from_stream, tag_status_ok, ImapFetchedHeader,
 };
-
-pub(crate) fn imap_probe(
-    host: &str,
-    port: u16,
-    use_tls: bool,
-    username: &str,
-    password: &str,
-    include_list: bool,
-) -> std::result::Result<(String, Vec<String>, Vec<String>), String> {
-    if password.is_empty() {
-        return Err("Missing credential secretValue on external account".to_string());
-    }
-
-    let addr = (host, port)
-        .to_socket_addrs()
-        .map_err(|e| format!("resolve failed: {e}"))?
-        .next()
-        .ok_or_else(|| "resolve failed: no address".to_string())?;
-
-    let tcp = TcpStream::connect_timeout(&addr, Duration::from_secs(10))
-        .map_err(|e| format!("tcp connect failed: {e}"))?;
-    tcp.set_read_timeout(Some(Duration::from_secs(12))).ok();
-    tcp.set_write_timeout(Some(Duration::from_secs(12))).ok();
-
-    if use_tls {
-        let connector = SslConnector::builder(SslMethod::tls())
-            .map_err(|e| format!("tls builder failed: {e}"))?
-            .build();
-        let ssl = connector
-            .connect(host, tcp)
-            .map_err(|e| format!("tls connect failed: {e}"))?;
-        run_imap_dialog_ssl(ssl, username, password, include_list)
-    } else {
-        run_imap_dialog_plain(tcp, username, password, include_list)
-    }
-}
 
 pub(crate) fn run_imap_dialog_plain(
     mut stream: TcpStream,
